+++ src-model/gpt_sovits_cpufast/__init__.py
@@ -1,3 +1,3 @@
-from gpt_sovits_v2pp import load_tts_params, load_voice_clone_params
+from .params import load_tts_params, load_voice_clone_params
 
 __all__ = ["load_tts_params", "load_voice_clone_params"]
