--- src-model/gpt_sovits_v2pp/__init__.py
@@ -1,3 +1,0 @@
-from .params import load_tts_params, load_voice_clone_params
-
-__all__ = ["load_tts_params", "load_voice_clone_params"]
