--- src-model/gpt_sovits_v2pp/requirements.txt
@@ -1,2 +1,0 @@
-# GPT-SoVITS-V2PP currently shells out to an external GPT-SoVITS-CPUFast checkout.
-# Keep this file intentionally minimal until runtime bootstrap/install logic is added.
