+++ src-model/scripts/windows/package_src_model.ps1
@@ -41,7 +41,7 @@ function Get-RelativeArchivePath {
 $excludeDirectoryNames = @('base-models', 'tests', '__pycache__', '.pytest_cache', '.mypy_cache', '.ruff_cache', 'venv', '.venv')
 $excludeExtensions = @('.pyc', '.pyo')
 $sourceDirectories = @('scripts', 'configs')
-$modelDirectories = @('qwen3_tts', 'vox_cpm2', 'moss_tts_local', 'gpt_sovits_v2pp', 'gpt_sovits_cpufast')
+$modelDirectories = @('qwen3_tts', 'vox_cpm2', 'moss_tts_local', 'gpt_sovits_cpufast')
 
 function Test-ShouldExcludeArchivePath {
     param(
