+++ src/views/VoiceCloneView.vue
@@ -77,7 +77,7 @@ interface SelectedAudioFile {
   filePath: string;
 }
 
-const DYNAMIC_REFERENCE_BASE_MODELS = new Set(['gpt_sovits_v2pp', 'gpt_sovits_cpufast']);
+const DYNAMIC_REFERENCE_BASE_MODELS = new Set(['gpt_sovits_cpufast']);
 const DEFAULT_EXPORT_AUDIO_NAME = createTaskExportAudioName(HistoryTaskType.VoiceClone);
 
 const uiConfigStore = useUiConfigStore();
